+++ src/chart/view.js
@@ -41,6 +41,16 @@ function isPointInCoord(coord, point) {
   return result;
 }
 
+const ViewGeoms = {};
+Util.each(Geom, function(geomConstructor, className) {
+  const methodName = Util.lowerFirst(className);
+  ViewGeoms[methodName] = function(cfg) {
+    const geom = new geomConstructor(cfg);
+    this.addGeom(geom);
+    return geom;
+  };
+});
+
 /**
  * 图表中的视图
  * @class View
@@ -83,6 +93,7 @@ class View extends Base {
     // Util.mix(this, ViewGeoms);
     this.init();
   }
+
 
   /**
    * @protected
@@ -323,6 +334,7 @@ class View extends Base {
     const guideController = this.get('guideController');
     if (!Util.isEmpty(guideController.options)) {
       const coord = this.get('coord');
+      guideController.view = this;
       guideController.backContainer = this.get('backPlot');
       guideController.frontContainer = this.get('frontPlot');
       guideController.xScales = this._getScales('x');